feat(admin): add upload and delete functionality for songs and albums

+++ backend/src/controllers/admin.controller.js
@@ -16,7 +16,7 @@ const uploadToCloudinary = async (file) => {
 
 export const createSong = async (req, res, next) => {
     try {
-        if (!req.files || !req.files.audioFile || !req.imageFile) {
+        if (!req.files || !req.files.audioFile || !req.files.imageFile) {
             return res.status(400).json({ message: "Please upload all files" })
         }
         const { title, artist, albumId, duration } = req.body;
@@ -31,7 +31,8 @@ export const createSong = async (req, res, next) => {
             artist,
             imageUrl,
             audioUrl,
-            albumId: albumId || null
+            albumId: albumId || null,
+            duration
         });
         await song.save();
 
@@ -51,11 +52,14 @@ export const deleteSong = async (req, res, next) => {
     const { id } = req.params;
     const song = await Song.findById(id);
 
-    if(song.albumId) {
-        await Album.findByIdAndDelete(song.albumId, {
-            $pull: {songs: song._id}
-        })
+    if (!song) return res.status(404).json({ message: "Song not found" });
+
+    if (song.albumId) {
+        await Album.findByIdAndUpdate(song.albumId, {
+            $pull: { songs: song._id }
+        });
     }
+
 
     await Song.findByIdAndDelete(id);
     res.status(200).json({ message: "Song deleted successfully" });
@@ -63,14 +67,14 @@ export const deleteSong = async (req, res, next) => {
 
 export const createAlbum = async (req, res, next) => {
     try {
-        if (!req.files || !req.files.audioFile || !req.imageFile) {
+        if (!req.files || !req.files.imageFile) {
             return res.status(400).json({ message: "Please upload all files" })
         }
+
         const { title, artist, releaseYear } = req.body;
         const { imageFile } = req.files;
 
         const imageUrl = await uploadToCloudinary(imageFile);
-
 
         const album = new Album({
             title,
@@ -90,9 +94,9 @@ export const createAlbum = async (req, res, next) => {
 export const deleteAlbum = async (req, res, next) => {
     try {
         const { id } = req.params;
-    await Song.deleteMany({ albumId: id });
-    await Album.findByIdAndDelete(id);
-    res.status(200).json({ message: "Album deleted successfully!" });
+        await Song.deleteMany({ albumId: id });
+        await Album.findByIdAndDelete(id);
+        res.status(200).json({ message: "Album deleted successfully!" });
     } catch (error) {
         console.log("Error in deleteAlbum", error);
         next(error);
